+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Support Bank of Alexandria (AlexBank)
 - Support Blom Bank Egypt (Blom)
 - Support Abu Dhabi Islamic Bank (ADIB)
+- Support Egyptian Gulf Bank (EGB)
 - Fix BanqueMisr swapping CHF and SEK exchange rates
 
 # v0.3.0