+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support Blom Bank Egypt (Blom)
 - Support Abu Dhabi Islamic Bank (ADIB)
 - Support Egyptian Gulf Bank (EGB)
+- Support National Bank of Greece (NBG)
 - Fix BanqueMisr swapping CHF and SEK exchange rates
 
 # v0.3.0