+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support Abu Dhabi Islamic Bank (ADIB)
 - Support Egyptian Gulf Bank (EGB)
 - Support National Bank of Greece (NBG)
+- Add API methods to easily interact with the gem.
 - Fix BanqueMisr swapping CHF and SEK exchange rates
 
 # v0.3.0